Conflicts:
	src/main/java/com/markany/blinkist/jsoup/HumanitiesCrawling.java
	src/main/java/com/markany/blinkist/jsoup/NovelCrawling.java
	src/main/java/com/markany/blinkist/jsoup/livingCrawling.java

+++ src/main/java/com/markany/blinkist/vo/ContentVo.java
@@ -4,10 +4,10 @@ import lombok.Data;
 
 @Data
 public class ContentVo {
-	private long content_no;
+	private int content_no;
 	private String chapter_title;
 	private String content;
 	private int chapter_no;
 	private String voice_path;
-	private long book_no;
+	private int book_no;
 }
